requirements.txt
torch==2.3.0
accelerate @ git+https://github.com/huggingface/accelerate.git
bitsandbytes
datasets
transformers==4.31
peft @ git+https://github.com/huggingface/peft.git
trl @ git+https://github.com/lvwerra/trl.git
scipy
wandb
SentencePiece
tensorflow-gpu==2.8.0
python-dotenv
